# docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noor API documentation",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "https://wwwwwwww123.herokuapp.com/",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "access-token",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/location/create": {
      "post": {
        "tags": [
          "Location"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "example": "any"
                },
                "address_line1": {
                  "example": "any"
                },
                "address_line2": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "postal_code": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/location/update/{id}": {
      "post": {
        "tags": [
          "Location"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "example": "any"
                },
                "address_line1": {
                  "example": "any"
                },
                "address_line2": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "postal_code": {
                  "example": "any"
                },
                "country": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/location/delete/{id}": {
      "delete": {
        "tags": [
          "Location"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/location/getone/{id}": {
      "get": {
        "tags": [
          "Location"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Xatolik roy berdi"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "Doston"
        },
        "last_name": {
          "type": "string",
          "example": "Ashurov"
        },
        "father_name": {
          "type": "string",
          "example": "Baxtiyorovich"
        },
        "email": {
          "type": "string",
          "example": "ashurov@list.ru"
        },
        "img_id": {
          "type": "string",
          "example": "6184b677f552e1ab7baf6fa2"
        },
        "phone": {
          "type": "string",
          "example": "94-456-7891"
        },
        "password": {
          "type": "string",
          "example": "789456qwer#A"
        },
        "role": {
          "type": "string",
          "example": "admin"
        }
      }
    },
...
  }
}